Moved some things around and more
Moved things around, and also fixed the hex colors. (Any hex color works now as before it wouldnt)
